+++ tests/dummy/app/routes/examples/infinite-scroll/controller.js
@@ -10,18 +10,18 @@ export default Controller.extend({
   actions: {
 
     loadAbove() {
-      let first = this.get('model.first');
+      let first = this.model.first;
       let numbers = getNumbers(first - 20, 20);
-      let model = this.get('model.numbers');
+      let model = this.model.numbers;
       model.unshiftObjects(numbers);
       // this.set('model.numbers', newModel);
       this.set('model.first', first - 20);
     },
 
     loadBelow() {
-      let last = this.get('model.last');
+      let last = this.model.last;
       let numbers = getNumbers(last, 20);
-      let model = this.get('model.numbers');
+      let model = this.model.numbers;
       model.pushObjects(numbers);
       // this.set('model.numbers', newModel);
       this.set('model.last', last + 20);